Made amount handle decimals

+++ server/server/Models/Expense.cs
@@ -2,7 +2,7 @@
 {
     public class Expense
     {
-        public int amount { get; set; }
+        public double amount { get; set; }
 
         public string category { get; set; }
     }
